feat(sx)!: smart autojump-to-first-match
BREAKING CHANGE: API changes and changed defaults in `opts`.

Auto-jump to first match is an implicit, per sublist decision now,
based on a dead simple heuristic: the probability of the user aiming for
the very first match lessens with the number of matches.

`opts.labels` has been divided into a `labels` and a `safe_labels` list,
where the latter is assumed to contain characters that are unlikely to
be used right after a jump. When the number of matches does not exceed
the number of safe labels (plus one), we will automatically jump to the
first match, otherwise stay in place and revert to the default list.

The flag `jump_to_first_match` has been removed; keeping a configuration
option to enforce auto-jumping in all circumstances does not seem
reasonable, as the gains of being able to use any key as a label when
there are lots of matches far outweigh the inconvenience of occasionally
having to use a label for the first match too.

See :h lightspeed-config for further details.

Also: some non-related tweaks and corrections in the help file.